Update package.py
check for file existence before fail-check

+++ package.py
@@ -314,6 +314,8 @@ class ZipWriteStream:
     def close(self, failed=False):
         self._zip.close()
         self._zip = None
+        if not os.exists(self._tmp_filename):
+            return        
         if failed:
             os.unlink(self._tmp_filename)
         else:
